+++ src/SyntaxBuilders/MigrationSyntaxBuilder.php
@@ -26,8 +26,6 @@ class MigrationSyntaxBuilder extends SyntaxBuilder
     {
         $syntax = sprintf("\$table->%s('%s')", $field['type'], $field['name']);
 
-        // If there are arguments for the schema type, like decimal('amount', 5, 2)
-        // then we have to remember to work those in.
         if ($field['arguments']) {
             $syntax = substr($syntax, 0, -1) . ', ';