+++ src/SyntaxBuilders/ViewIndexSyntaxBuilder.php
@@ -1,7 +1,6 @@
 <?php
 
 namespace FaceDigital\Crudify\SyntaxBuilders;
-
 use FaceDigital\Crudify\Commands\Traits\NamingConvention;
 use Illuminate\Support\Str;
 
@@ -42,13 +41,23 @@ class ViewIndexSyntaxBuilder extends SyntaxBuilder
 
     private function addColumn(array $field): string
     {
-        return sprintf("<th scope='col'>%s</th>", Str::studly($field['name']));
-        // return sprintf("<div class='px-4 py-2'>%s</div>", Str::studly($field['name']));
+        if(config('crudify.theme') == 'bootstrap') {
+            return sprintf("<th scope='col'>%s</th>", Str::studly($field['name']));
+        }
+
+        if(config('crudify.theme') == 'tailwind') {
+            return sprintf("<div class='px-4 py-2'>%s</div>", Str::studly($field['name']));
+        }
     }
 
     private function addRow(array $field): string
     {
-        // return sprintf("<div class='px-4 py-2'>{{ $%s->%s }}</div>", '{{singularName}}', $field['name']);
-        return sprintf("<td>{{ $%s->%s }}</td>", '{{singularName}}', $field['name']);
+        if(config('crudify.theme') == 'bootstrap') {
+            return sprintf("<td>{{ $%s->%s }}</td>", '{{singularName}}', $field['name']);
+        }
+
+        if(config('crudify.theme') == 'tailwind') {
+            return sprintf("<div class='px-4 py-2'>{{ $%s->%s }}</div>", '{{singularName}}', $field['name']);
+        }
     }
 }